Move ifex/model to models/ifex [newstructure]
