+++ ifex/schema/ifex_to_json_schema.py
@@ -16,10 +16,10 @@ Generate JSON Schema equivalent to the python-internal model definition.
 from dataclasses import fields
 from datetime import datetime, date
 
-from ifex.model.ifex_ast_construction import is_simple_type
+from models.ifex.ifex_ast_construction import is_simple_type
 
-from ifex.model.ifex_ast import AST
-from ifex.model.ifex_ast_introspect import actual_type, field_actual_type, field_inner_type, inner_type, is_forwardref, is_list, type_name, field_is_list, field_is_optional, field_referenced_type
+from models.ifex.ifex_ast import AST
+from models.ifex.ifex_ast_introspect import actual_type, field_actual_type, field_inner_type, inner_type, is_forwardref, is_list, type_name, field_is_list, field_is_optional, field_referenced_type
 from typing import Any
 import json
